Small refactor for bettter organization
