+++ datamint/dataset/sliced_dataset.py
@@ -4,22 +4,15 @@ SlicedVolumeDataset - 2D dataset created by slicing a VolumeDataset along an axi
 Provides a way to iterate over individual 2D slices from 3D volume data,
 enabling training of 2D models on volumetric medical imaging data.
 """
-import gzip
+from __future__ import annotations
 import hashlib
-import logging
-from functools import cached_property
 from typing import Any, Literal, TYPE_CHECKING
 from typing_extensions import override
 from collections.abc import Sequence
-
 import numpy as np
 import torch
 from torch import Tensor
 import albumentations
-
-from medimgkit.readers import read_array_normalized
-from medimgkit import dicom_utils
-from medimgkit import nifti_utils
 
 from .base import DatamintBaseDataset
 from .annotation_processor import AnnotationProcessor, MergeStrategy
@@ -27,195 +20,13 @@ from .annotation_processor import AnnotationProcessor, MergeStrategy
 from datamint.entities.cache_manager import CacheManager
 if TYPE_CHECKING:
     from datamint.entities import Annotation, Resource
-
-_LOGGER = logging.getLogger(__name__)
-
-# Cache key for parsed slice numpy arrays
-_SLICE_ARRAY_CACHEKEY = "slice_array"
+    from datamint.entities.sliced_resource import SlicedVolumeResource
 
 # Cache key for sliced segmentation numpy arrays
 _SEG_SLICE_CACHEKEY = "seg_slice_array"
 
 
-class SlicedVolumeResource:
-    """Proxy that presents a single 2D slice of a 3D volume Resource.
 
-    This class wraps a :class:`Resource` and represents a specific 2D slice
-    along a given axis. It uses gzip-compressed ``.npy.gz`` files on disk
-    for efficient storage, with an in-memory LRU cache managed by
-    :class:`CacheManager`.
-
-    The CacheManager memory cache is disabled by default globally, but the
-    sliced-volume cache manager enables it by default.
-
-    This shared cache avoids repeated gzip decompression for already-cached
-    slices. Full-volume caching is intentionally not handled here.
-
-    Args:
-        parent: The original 3D volume Resource.
-        slice_index: The index of the slice along the given axis.
-        slice_axis: The spatial axis to slice along ('axial', 'coronal', 'sagittal').
-        sliced_vols_cache: Shared :class:`CacheManager` for disk-based volume caching.
-    """
-
-    _CACHE_MANAGER_NAMESPACE = "sliced_volumes"
-
-    def __init__(
-        self,
-        parent: 'Resource',
-        slice_index: int,
-        slice_axis: str,
-        sliced_vols_cache: CacheManager | None = None,
-    ):
-        self._parent = parent
-        self.slice_index = slice_index
-        self.slice_axis = slice_axis
-        if sliced_vols_cache is None:
-            sliced_vols_cache = CacheManager(SlicedVolumeResource._CACHE_MANAGER_NAMESPACE)
-        self._volume_cache = sliced_vols_cache
-
-    @staticmethod
-    def slice_over(resource: 'Resource',
-                   slice_axis: Literal['axial', 'coronal', 'sagittal'],
-                   volume_cache: CacheManager | None = None) -> list['SlicedVolumeResource']:
-        sliced_resources = []
-        res_data = resource.fetch_file_data(auto_convert=True, use_cache=True)
-        if resource.is_dicom():
-            axis_size = dicom_utils.get_dim_size(res_data, slice_axis)
-        elif resource.is_nifti():
-            axis_size = nifti_utils.get_dim_size(res_data, slice_axis)
-        else:
-            raise ValueError(f"Unsupported resource type for slicing axis: {resource.filename}|{resource.mimetype}")
-
-        # anns = resource_annotations[i]
-
-        for s in range(axis_size):
-            sliced_resources.append(
-                SlicedVolumeResource(resource, s, slice_axis, volume_cache)
-            )
-
-        return sliced_resources
-
-    def __getattr__(self, name: str) -> Any:
-        """Delegate all unresolved attributes to the parent Resource."""
-        return getattr(self._parent, name)
-
-    def get_depth(self) -> int:
-        """A single slice has depth 1."""
-        return 1
-
-    def _get_version_info(self) -> dict:
-        """Get version info from the parent resource for cache validation."""
-        return {
-            'created_at': self._parent.created_at,
-            'deleted_at': self._parent.deleted_at,
-            'size': self._parent.size,
-        }
-
-    def _slice_cache_entity_id(self) -> str:
-        return f"{self._parent.id}:axis{self.slice_axis}:slice{self.slice_index}"
-
-    def fetch_slice_data(self) -> np.ndarray:
-        """Fetch the 2D slice as a (C, H, W) array.
-
-        Returns:
-            Slice array with shape (C, H, W).
-        """
-        version_info = self._get_version_info()
-        cache_entity_id = self._slice_cache_entity_id()
-
-        cached_slice = self._volume_cache.get(
-            cache_entity_id,
-            _SLICE_ARRAY_CACHEKEY,
-            version_info,
-        )
-        if cached_slice is not None:
-            return np.ascontiguousarray(cached_slice)
-
-        raw = self._parent.fetch_file_data(auto_convert=False, use_cache=True)
-        vol, self.data_metainfo = read_array_normalized(raw, return_metainfo=True)
-        # vol.shape is (D,C,H,W)
-
-        _LOGGER.debug(
-            f"Slicing {self._parent.filename} along axis {self.slice_axis=} ({self.slice_axis_idx_std=}) at index {self.slice_index=} with shape {vol.shape=}")
-        sliced = np.take(vol, self.slice_index, axis=self.slice_axis_idx_std)
-        # vol is (D, C, H, W); after np.take the sliced axis is removed.
-        # C was at index 1. If we sliced axis 0 (D), C shifts to index 0 — already correct.
-        # If we sliced axis 2 (H) or 3 (W), C stays at index 1 → move it to 0.
-        channel_axis_in_result = 1 if self.slice_axis_idx_std > 1 else 0
-        if channel_axis_in_result != 0:
-            sliced = np.moveaxis(sliced, channel_axis_in_result, 0)
-        # sliced is now (C, DIM1, DIM2)
-        sliced = np.ascontiguousarray(sliced)
-
-        gz_path = self._volume_cache.get_expected_path(cache_entity_id, _SLICE_ARRAY_CACHEKEY)
-        gz_path = gz_path.with_suffix('.npy.gz')
-        gz_path.parent.mkdir(parents=True, exist_ok=True)
-        with gzip.open(str(gz_path), 'wb', compresslevel=4) as f:
-            np.save(f, sliced)
-
-        self._volume_cache.register_file_location(
-            cache_entity_id,
-            _SLICE_ARRAY_CACHEKEY,
-            file_path=gz_path,
-            version_info=version_info,
-            mimetype='application/gzip',
-            data=sliced,
-        )
-
-        _LOGGER.debug(f'Sliced shape: {sliced.shape}')
-
-        return sliced
-
-    @cached_property
-    def data_metainfo(self) -> dict:
-        """Volume metadata. Loaded once and cached for the lifetime of this resource."""
-        raw = self._parent.fetch_file_data(auto_convert=False, use_cache=True)
-        _, metainfo = read_array_normalized(raw, return_metainfo=True)
-        return metainfo
-
-    @cached_property
-    def slice_axis_idx(self) -> int:
-        """Raw axis index for the slice axis. Computed once and cached."""
-        if self._parent.is_dicom():
-            return dicom_utils.get_plane_axis(self.data_metainfo, self.slice_axis)
-        elif self._parent.is_nifti():
-            return nifti_utils.get_plane_axis(self.data_metainfo, self.slice_axis)
-        else:
-            raise ValueError(
-                "Unsupported resource type for slicing axis:"
-                f" {self._parent.filename}|{self._parent.mimetype}|{self.slice_axis}")
-
-    @cached_property
-    def slice_axis_idx_std(self) -> int:
-        """Standardised axis index for the slice axis. Computed once and cached."""
-        if self._parent.is_dicom():
-            return dicom_utils.rawplaneaxis2stdplaneaxis_idx(self.slice_axis_idx)
-        elif self._parent.is_nifti():
-            return nifti_utils.rawplaneaxis2stdplaneaxis_idx(self.slice_axis_idx)
-        else:
-            raise ValueError(
-                f"Unsupported resource type for slicing axis: {self._parent.filename}|{self._parent.mimetype}")
-
-    @property
-    def parent_resource(self) -> 'Resource':
-        """The original volume Resource being proxied."""
-        return self._parent
-
-    def __repr__(self) -> str:
-        axis_names = {0: 'axial', 1: 'coronal', 2: 'sagittal'}
-        axis_name = axis_names.get(self.slice_axis, str(self.slice_axis))
-        return (
-            f"SlicedVolumeResource(filename='{self._parent.filename}', "
-            f"axis='{axis_name}', slice={self.slice_index})"
-        )
-
-    def __getattribute__(self, name: str) -> Any:
-        try:
-            return super().__getattribute__(name)
-        except AttributeError:
-            parent = super().__getattribute__('_parent')
-            return getattr(parent, name)
 
 
 # # Axis mapping for anatomical orientations
@@ -247,7 +58,7 @@ class SlicedVolumeDataset(DatamintBaseDataset):
 
     def __init__(
         self,
-        parent_dataset: 'DatamintBaseDataset',
+        parent_dataset: DatamintBaseDataset,
         slice_axis: Literal['axial', 'coronal', 'sagittal'] | int = 'axial',
     ):
         # We intentionally do NOT call super().__init__() because that
@@ -339,6 +150,8 @@ class SlicedVolumeDataset(DatamintBaseDataset):
         Returns:
             Tuple of (sliced_resources, sliced_annotations).
         """
+        from datamint.entities.sliced_resource import SlicedVolumeResource
+        
         sliced_resources: list[SlicedVolumeResource] = []
         sliced_annotations: list[Sequence['Annotation']] = []
